Optimized auto creation of groups
